+++ GrowDiary.Web/Program.cs
@@ -76,6 +76,9 @@ builder.Services.AddScoped<SensorReadingRepository>();
 builder.Services.AddScoped<AutoMeasurementExecutionService>();
 builder.Services.AddScoped<AlertRuleRepository>();
 builder.Services.AddScoped<AlertEvaluationService>();
+builder.Services.AddSingleton<NotificationSettingsRepository>();
+builder.Services.AddScoped<NotificationService>();
+builder.Services.AddScoped<CalibrationReminderService>();
 builder.Services.AddHostedService<HomeAssistantSnapshotWorker>();
 builder.Services.AddHostedService<AutoMeasurementWorker>();